服を表示するVCに継承を用いる（仮実装)
